Update starchiverT3-ext.sh
Added - New parameter --cleanup
This parameter will perform a scan for Obsolete Hashes and build a list.  The move all of the Obsolete Hashes into a folder called Obsolete_Hashes.  Which then the user can choose to remove to reclaim disk space.

Added - A new option in the main menu
O - Cleanup Options
The Cleanup Options allow the user to choose to perform a scan for Obsolete Hashes or if an Obsolete_Hashes folder is found in the same directory as the script, it will allow the user to choose to Reclaim Disk Space, by removing all of the Obsolete Hashes that have been removed from a scan.

Fixed - Various other fixes and color code was added