Update starchiverT3-ext.sh
Added: The ability for Starchive Extractor to detect if the server it is launched on has any ordinal files that are missing from the Starchiver sets.  If it detects them, it will give the user a message asking if they would let Starchiver compile them into packages and upload them. Then the user can share the urls it creates with Proph151Music.

Added: Implemented tmux to put Starchiver into it's own terminal session.  So that the Starchiver session can be detached and reattached for viewing.  This will also allow Starchiver to run even if the user logs off the server or loses connection.  If the user is detached and the session is still running, they can get back to the session by simply re-launching Starchiver again.

Added other improvements throughout the code.