Fix --datetime ordinal
Fix - When specifying an ordinal with --datetime, If the ordinal specified was higher then the ordinal set number of the latest ordinal set, Starchiver would incorrectly state that the ordinal was newer then the latest set.  Even if the ordinal fell within that set according to the -e of the latest ordinal set archive name.